backend/requirements.txt
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0
torch==2.1.0
torchvision==0.16.0
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6
joblib==1.3.0
matplotlib==3.7.0
pytesseract==0.3.10
Pillow==10.0.0
python-dotenv==1.0.0
supabase==2.0.0
pydantic==2.4.0
opencv-python==4.8.0.76
